chore: Separate modules for resource backend
Reduce code size of the sql.rs file by splitting the individual resource
into separate modules.
